backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.4
alembic==1.14.0
python-dotenv==1.0.1
aiofiles==24.1.0
